requirements.txt
notion-client
slack-sdk
python-dotenv